Tweak font-sizes
- Use 'rem' for body size
- Use 'em' values for everything
- Use 14px for mobile body size
- Explicitly set the section li font-size
  Otherwise li's not inside <section> will be sized oddly

When it comes to fonts generally you want the "global" body font-size
to be a 'rem' or 'px' value, and the others 'em' values (which will
scale based on that global font-size). This way, when for example on
mobile  the font-size is different, all the other elements will scale
accordingly.